Replace type checking with stronger HTML check
Asserting that we have a Hash doesn't check the structure of the output.
We also don't care that it is a Hash, but just that it behaves like a Hash.

The negative assertion (not Hash) would have failed in a situation where
we had something that would work like a Hash just fine.

[#90957612]